+++ src/components/common/form/_jss/PasswordField.jss.ts
@@ -1,13 +1,6 @@
-import { Theme, makeStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 
-export default makeStyles((theme: Theme) => ({
-  formikTextField: {
-    marginBottom: theme.spacing(0.5),
-  },
-  formikTextFieldHelper: {
-    marginLeft: 4,
-    marginTop: 0,
-  },
+export default makeStyles(() => ({
   inputContainer: {
     overflow: 'hidden',
   },
