+++ src/components/common/form/_jss/TextField.jss.ts
@@ -1,0 +1,11 @@
+import { Theme, makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles((theme: Theme) => ({
+  formikTextField: {
+    marginBottom: theme.spacing(0.5),
+  },
+  formikTextFieldHelper: {
+    marginLeft: 4,
+    marginTop: 0,
+  },
+}));
